requirements.txt
psycopg2-binary==2.9.9
sentence-transformers==2.7.0
langchain
langchain-community
langchain-ollama
pgvector==0.2.5
python-dotenv==1.0.1
ollama==0.5.1
SQLAlchemy==2.0.30
gradio==4.29.0
